endian fixes
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@42492 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/smbd/smbd/smb1pdu.h
@@ -1665,7 +1665,7 @@ extern int init_smb1_server(struct ksmbd_conn *conn);
 
 /* function prototypes */
 extern int init_smb_rsp_hdr(struct ksmbd_work *work);
-extern int get_smb_cmd_val(struct ksmbd_work *work);
+extern uint16_t get_smb_cmd_val(struct ksmbd_work *work);
 extern void set_smb_rsp_status(struct ksmbd_work *work, __le32 err);
 extern int smb_allocate_rsp_buf(struct ksmbd_work *work);
 extern int smb1_is_sign_req(struct ksmbd_work *work, unsigned int command);
